[FIX] ir.model.data: ensure cache of known "model data reference ids" is not lost after osv inheritance
By storing this cache in the pool and only referencing it in the instance we ensure that it is not lost
when the osv instance is replaced in the pool, e.g. as soon as on osv object inherits from ir.model.data.
This is now needed, as the module data cleanup is now done globally at the end of all module installation,
while it was done on a smaller scale before (leading to other kinds of issues).