[FIX] res.users: proper multicompany record rule for users, courtesy of Valentin Lab
Contrary to regular records, the company_id field
on res_users does not represent a permanent
relationship, this field is rather used by the
user to select the 'company context' in which
she would like to work.
We don't want the list of assignable users on
a task to change depending on which company
these users are currently viewing data, for
example.
The real field that holds the list of companies
a user belongs to is the 'company_ids' field.